+++ lib/sonos-http-api.js
@@ -3,6 +3,7 @@ const requireDir = require('./helpers/require-dir');
 const path = require('path');
 const request = require('sonos-discovery/lib/helpers/request');
 const logger = require('sonos-discovery/lib/helpers/logger');
+const HttpEventServer = require('./helpers/http-event-server');
 
 function HttpAPI(discovery, settings) {
 
@@ -163,24 +164,6 @@ function HttpAPI(discovery, settings) {
     })
   }
 
-}
-
-function HttpEventServer() {
-  let clients = [];
-
-  const removeClient = client => clients = clients.filter(value => value !== client);
-
-  this.addClient = res => clients.push(new HttpEventSource(res, removeClient));
-
-  this.sendEvent = event => clients.forEach(client => client.sendEvent(event))
-}
-
-function HttpEventSource(res, done) {
-  this.sendEvent = event => res.write('data: ' + event + '\n\n')
-
-  res.on('close', () => done(this))
-
-  res.setHeader('Content-Type', 'text/event-stream');
 }
 
 module.exports = HttpAPI;
